resolve nats code climate issues
