+++ pkg/gofr/datasource/pubsub/nats/client.go
@@ -51,7 +51,7 @@ type messageHandler func(context.Context, jetstream.Msg) error
 func (c *Client) Connect() error {
 	c.logger.Debugf("connecting to NATS server at %v", c.Config.Server)
 
-	if err := c.validateAndPrepare(); err != nil {
+	if err := validateAndPrepare(c.Config, c.logger); err != nil {
 		return err
 	}
 
@@ -64,59 +64,6 @@ func (c *Client) Connect() error {
 	}
 
 	return nil
-}
-
-// establishConnection handles the actual connection process to NATS and sets up jStream.
-func (c *Client) establishConnection() error {
-	connManager := NewConnectionManager(c.Config, c.logger, c.natsConnector, c.jetStreamCreator)
-	if err := connManager.Connect(); err != nil {
-		return err
-	}
-
-	c.connManager = connManager
-
-	js, err := c.connManager.jetStream()
-	if err != nil {
-		return err
-	}
-
-	c.streamManager = newStreamManager(js, c.logger)
-	c.subManager = newSubscriptionManager(batchSize)
-	c.logSuccessfulConnection()
-
-	return nil
-}
-
-func (c *Client) retryConnect() {
-	for {
-		c.logger.Debugf("connecting to NATS server at %v", c.Config.Server)
-
-		if err := c.establishConnection(); err != nil {
-			c.logger.Errorf("failed to connect to NATS server at %v: %v", c.Config.Server, err)
-
-			time.Sleep(defaultRetryTimeout)
-
-			continue
-		}
-
-		return
-	}
-}
-
-func (c *Client) validateAndPrepare() error {
-	if err := validateConfigs(c.Config); err != nil {
-		c.logger.Errorf("could not initialize NATS jStream: %v", err)
-
-		return err
-	}
-
-	return nil
-}
-
-func (c *Client) logSuccessfulConnection() {
-	if c.logger != nil {
-		c.logger.Logf("connected to NATS server '%s'", c.Config.Server)
-	}
 }
 
 // UseLogger sets the logger for the NATS client.
@@ -169,10 +116,6 @@ func (c *Client) Subscribe(ctx context.Context, topic string) (*pubsub.Message, 
 	}
 }
 
-func (c *Client) generateConsumerName(subject string) string {
-	return fmt.Sprintf("%s_%s", c.Config.Consumer, strings.ReplaceAll(subject, ".", "_"))
-}
-
 func (c *Client) SubscribeWithHandler(ctx context.Context, subject string, handler messageHandler) error {
 	c.subMutex.Lock()
 	defer c.subMutex.Unlock()
@@ -185,7 +128,7 @@ func (c *Client) SubscribeWithHandler(ctx context.Context, subject string, handl
 		return err
 	}
 
-	consumerName := c.generateConsumerName(subject)
+	consumerName := fmt.Sprintf("%s_%s", c.Config.Consumer, strings.ReplaceAll(subject, ".", "_"))
 
 	cons, err := c.createOrUpdateConsumer(ctx, js, subject, consumerName)
 	if err != nil {
@@ -211,81 +154,6 @@ func (c *Client) cancelExistingSubscription(subject string) {
 	}
 }
 
-func (c *Client) createOrUpdateConsumer(
-	ctx context.Context, js jetstream.JetStream, subject, consumerName string) (jetstream.Consumer, error) {
-	cons, err := js.CreateOrUpdateConsumer(ctx, c.Config.Stream.Stream, jetstream.ConsumerConfig{
-		Durable:       consumerName,
-		AckPolicy:     jetstream.AckExplicitPolicy,
-		FilterSubject: subject,
-		MaxDeliver:    c.Config.Stream.MaxDeliver,
-		DeliverPolicy: jetstream.DeliverNewPolicy,
-	})
-	if err != nil {
-		c.logger.Errorf("failed to create or update consumer: %v", err)
-		return nil, err
-	}
-
-	return cons, nil
-}
-
-func (c *Client) processMessages(ctx context.Context, cons jetstream.Consumer, subject string, handler messageHandler) {
-	for ctx.Err() == nil {
-		if err := c.fetchAndProcessMessages(ctx, cons, subject, handler); err != nil {
-			c.logger.Errorf("Error in message processing loop for subject %s: %v", subject, err)
-		}
-	}
-}
-
-func (c *Client) fetchAndProcessMessages(ctx context.Context, cons jetstream.Consumer, subject string, handler messageHandler) error {
-	msgs, err := cons.Fetch(1, jetstream.FetchMaxWait(c.Config.MaxWait))
-	if err != nil {
-		if !errors.Is(err, context.DeadlineExceeded) {
-			c.logger.Errorf("Error fetching messages for subject %s: %v", subject, err)
-		}
-
-		return err
-	}
-
-	return c.processFetchedMessages(ctx, msgs, handler, subject)
-}
-
-func (c *Client) processFetchedMessages(ctx context.Context, msgs jetstream.MessageBatch, handler messageHandler, subject string) error {
-	for msg := range msgs.Messages() {
-		if err := c.handleMessage(ctx, msg, handler); err != nil {
-			c.logger.Errorf("Error processing message: %v", err)
-		}
-	}
-
-	if err := msgs.Error(); err != nil {
-		c.logger.Errorf("Error in message batch for subject %s: %v", subject, err)
-		return err
-	}
-
-	return nil
-}
-
-func (c *Client) handleMessage(ctx context.Context, msg jetstream.Msg, handler messageHandler) error {
-	err := handler(ctx, msg)
-	if err == nil {
-		if ackErr := msg.Ack(); ackErr != nil {
-			c.logger.Errorf("Error sending ACK for message: %v", ackErr)
-			return ackErr
-		}
-
-		return nil
-	}
-
-	c.logger.Errorf("Error handling message: %v", err)
-
-	if nakErr := msg.Nak(); nakErr != nil {
-		c.logger.Debugf("Error sending NAK for message: %v", nakErr)
-
-		return nakErr
-	}
-
-	return err
-}
-
 // Close closes the Client.
 func (c *Client) Close(ctx context.Context) error {
 	c.subManager.Close()
@@ -308,173 +176,29 @@ func (c *Client) Query(ctx context.Context, query string, args ...any) ([]byte, 
 	}
 
 	// Parse optional arguments
-	timeout, limit := c.parseQueryArgs(args...)
+	timeout, limit := parseQueryArgs(args...)
 
-	// Use provided context or add default timeout
-	queryCtx := ctx
-
-	if _, hasDeadline := ctx.Deadline(); !hasDeadline {
-		var cancel context.CancelFunc
-
-		queryCtx, cancel = context.WithTimeout(ctx, timeout)
-		defer cancel()
-	}
+	// Create a query context with timeout
+	queryCtx, cancel := c.createQueryContext(ctx, timeout)
+	defer cancel()
 
 	js, err := c.connManager.jetStream()
 	if err != nil {
 		return nil, err
 	}
 
-	streamName := c.Config.Stream.Stream
-	if query == goFrNatsStreamName {
-		streamName = goFrNatsStreamName
-	}
+	streamName := c.getStreamName(query)
+	consumerName := fmt.Sprintf("query_%s_%d", query, time.Now().UnixNano())
 
-	consumerName := fmt.Sprintf("query_%s_%d", c.generateConsumerName(query), time.Now().UnixNano())
-
-	// Configure consumer to read from beginning of stream
-	cons, err := js.CreateOrUpdateConsumer(queryCtx, streamName, jetstream.ConsumerConfig{
-		Durable:       consumerName,
-		AckPolicy:     jetstream.AckExplicitPolicy,
-		FilterSubject: query,
-		DeliverPolicy: jetstream.DeliverAllPolicy,
-		AckWait:       c.Config.MaxWait,
-	})
+	// Create a consumer
+	cons, err := c.createConsumer(queryCtx, js, streamName, query, consumerName)
 	if err != nil {
-		c.logger.Errorf("failed to create consumer for query: %v", err)
 		return nil, err
 	}
+	defer c.cleanupConsumer(js, streamName, cons)
 
-	defer func() {
-		// Clean up the temporary consumer
-		if deleteCtx, cancel := context.WithTimeout(context.Background(), defaultDeleteTimeout); cancel != nil {
-			defer cancel()
-
-			info, err := cons.Info(deleteCtx)
-
-			if err != nil {
-				c.logger.Debugf("failed to delete temporary consumer: %v", err)
-			}
-
-			// Delete consumer using JetStream interface
-			if err := js.DeleteConsumer(deleteCtx, streamName, info.Name); err != nil {
-				c.logger.Debugf("failed to delete temporary consumer: %v", err)
-			}
-		}
-	}()
-
-	return c.collectMessages(queryCtx, cons, limit)
-}
-
-// parseQueryArgs parses the query arguments.
-func (*Client) parseQueryArgs(args ...any) (timeout time.Duration, limit int) {
-	// Default values
-	timeout = defaultQueryTimeout
-	limit = 100
-
-	if len(args) > 0 {
-		// First argument can be a custom timeout
-		if val, ok := args[0].(time.Duration); ok && val > 0 {
-			timeout = val
-		}
-	}
-
-	if len(args) > 1 {
-		// Second argument is the message limit
-		if val, ok := args[1].(int); ok && val > 0 {
-			limit = val
-		}
-	}
-
-	return timeout, limit
-}
-
-// collectMessages fetches messages from the consumer and combines them.
-func (c *Client) collectMessages(ctx context.Context, cons jetstream.Consumer, limit int) ([]byte, error) {
-	var result []byte
-
-	messagesCollected := 0
-
-	for messagesCollected < limit {
-		// Fetch messages with a batch size based on remaining needed messages
-		fetchSize := minInt(batchSize, limit-messagesCollected)
-		if fetchSize <= 0 {
-			break
-		}
-
-		msgs, err := c.fetchBatch(cons, fetchSize)
-		if err != nil {
-			return result, err
-		}
-
-		collected := c.processBatch(ctx, msgs, &result, &messagesCollected, limit)
-		if !collected {
-			break
-		}
-	}
-
-	if strings.Contains(cons.CachedInfo().Config.FilterSubject, goFrNatsStreamName) && len(result) == 0 {
-		c.logger.Debugf("No migration records found in stream %s", c.Config.Stream.Stream)
-	}
-
-	return result, nil
-}
-
-func (c *Client) fetchBatch(cons jetstream.Consumer, fetchSize int) (jetstream.MessageBatch, error) {
-	msgs, err := cons.Fetch(fetchSize, jetstream.FetchMaxWait(c.Config.MaxWait))
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled) {
-			return nil, nil
-		}
-
-		c.logger.Errorf("Error fetching messages: %v", err)
-
-		return nil, err
-	}
-
-	return msgs, nil
-}
-
-func (c *Client) processBatch(ctx context.Context, msgs jetstream.MessageBatch,
-	result *[]byte, messagesCollected *int, limit int) bool {
-	receivedAny := false
-
-	for msg := range msgs.Messages() {
-		receivedAny = true
-
-		// Add newline separator between messages
-		if len(*result) > 0 {
-			*result = append(*result, '\n')
-		}
-
-		// Append message data
-		*result = append(*result, msg.Data()...)
-
-		// Acknowledge the message
-		if err := msg.Ack(); err != nil {
-			c.logger.Debugf("Error acknowledging message: %v", err)
-		}
-
-		*messagesCollected++
-		if *messagesCollected >= limit {
-			break
-		}
-	}
-
-	if !receivedAny || errors.Is(ctx.Err(), context.DeadlineExceeded) || errors.Is(ctx.Err(), context.Canceled) {
-		return false
-	}
-
-	return true
-}
-
-// Helper function for Go versions < 1.21.
-func minInt(a, b int) int {
-	if a < b {
-		return a
-	}
-
-	return b
+	// Collect messages
+	return collectMessages(queryCtx, cons, limit, c.Config, c.logger)
 }
 
 // CreateTopic creates a new topic (stream) in NATS jStream.
@@ -545,20 +269,4 @@ func GetJetStreamStatus(ctx context.Context, js jetstream.JetStream) (string, er
 	}
 
 	return jetStreamStatusOK, nil
-}
-
-func checkClient(c *Client) error {
-	if c == nil {
-		return errClientNotConnected
-	}
-
-	if c.connManager == nil {
-		return errClientNotConnected
-	}
-
-	if !c.connManager.isConnected() {
-		return errClientNotConnected
-	}
-
-	return nil
 }
